добавит разбивку на модули
